+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 
 ## UNRELEASED
 
+### Added
+- Tags can now be linked to a default assignee, who is preselected when the tag is added to a segment during statement splitting
+
 ## v4.50.0 (2026-07-15)
 
 ### Added